requirements.txt
streamlit
pandas
plotly
scikit-learn
Pillow

